prettier

+++ contracts/vault/VaultAdmin.sol
@@ -450,11 +450,11 @@ contract VaultAdmin is VaultStorage {
     {
         IStrategy strategy = IStrategy(_strategyAddr);
         _harvest(address(strategy));
-        address[] memory rewardTokenAddresses = strategy.getRewardTokenAddresses();
-        for (uint i = 0; i < rewardTokenAddresses.length; i++) {
+        address[] memory rewardTokenAddresses = strategy
+            .getRewardTokenAddresses();
+        for (uint256 i = 0; i < rewardTokenAddresses.length; i++) {
             _swap(rewardTokenAddresses[i]);
         }
-
     }
 
     /**
@@ -464,7 +464,8 @@ contract VaultAdmin is VaultStorage {
      */
     function _harvest(address _strategyAddr) internal {
         IStrategy strategy = IStrategy(_strategyAddr);
-        address[] memory rewardTokenAddresses = strategy.getRewardTokenAddresses();
+        address[] memory rewardTokenAddresses = strategy
+            .getRewardTokenAddresses();
         if (rewardTokenAddresses.length != 0) {
             strategy.collectRewardTokens();
         }
